bsp: tf-a: add TF-A recipe with OP-TEE support
Recipe is based on similar recipe from meta-arm. We can't use meta-arm
directly because it relies on own set of toolchains and we don't want
to switch tool chain.

We are building latest (on time of writing) master branch for TF-A
becuse it got RPI5 support very recently and there is no official
releases that support RPI5.

This patch also enabled OP-TEE boot via TF-A. As a result of a build,
it will generate armstub8-2712.bin file, which is BL31+OP-TEE
combined. It then can be written to SD card to enable OP-TEE support
on the board.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>